New version that deals with continuous values by providing a collision threshold instead of requiring discrete values.

+++ src/main/scala/org/apache/spark/ml/feature/MatrixAccumulator.scala
@@ -7,22 +7,22 @@ import org.apache.spark.util.AccumulatorV2
 /**
  * @author sramirez
  */
-class MatrixAccumulator(val rows: Int, val cols: Int) extends AccumulatorV2[Matrix[Long], Matrix[Long]] {
+class MatrixAccumulator(val rows: Int, val cols: Int) extends AccumulatorV2[Matrix[Double], Matrix[Double]] {
   
-  def this(m: Matrix[Long]) = {
+  def this(m: Matrix[Double]) = {
     this(m.rows, m.cols)
     this.accMatrix = m.copy
   }
 
-  private var accMatrix: Matrix[Long] = Matrix.zeros[Long](rows, cols)
+  private var accMatrix: Matrix[Double] = Matrix.zeros[Double](rows, cols)
   private var zero: Boolean = true
 
   def reset(): Unit = {
-    accMatrix = Matrix.zeros[Long](rows, cols)
+    accMatrix = Matrix.zeros[Double](rows, cols)
     zero = true
   }
 
-  def add(v: Matrix[Long]): Unit = {
+  def add(v: Matrix[Double]): Unit = {
     if(isZero) 
       zero = false
     accMatrix += v
@@ -30,9 +30,9 @@ class MatrixAccumulator(val rows: Int, val cols: Int) extends AccumulatorV2[Matr
   
   def isZero(): Boolean = zero
   
-  def merge(other: AccumulatorV2[Matrix[Long], Matrix[Long]]): Unit = accMatrix += other.value
+  def merge(other: AccumulatorV2[Matrix[Double], Matrix[Double]]): Unit = accMatrix += other.value
   
-  def value: Matrix[Long] = accMatrix
+  def value: Matrix[Double] = accMatrix
   
-  def copy(): AccumulatorV2[Matrix[Long], Matrix[Long]] = new MatrixAccumulator(accMatrix)
+  def copy(): AccumulatorV2[Matrix[Double], Matrix[Double]] = new MatrixAccumulator(accMatrix)
 }
